Cleaned up autoloads and global types removed
