+++ Engine/Autoloads/Multiplayer/Lobby/Utility.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using GodotPeer2PeerSteamCSharp.Types.Lobby;
 using GodotPeer2PeerSteamCSharp.Types.Scene;
 
 namespace GodotPeer2PeerSteamCSharp.Core.Lobby;
@@ -13,7 +14,7 @@ public partial class LobbyManager
         Logger.Lobby($"Lobby members gathered: {LobbyMembersData.Players.Count}");
     }
     
-    public static Task<List<GlobalTypes.PlayerInvite>> GetInGameFriends()
+    public static Task<List<PlayerInvite>> GetInGameFriends()
     {
         try
         {
@@ -22,7 +23,7 @@ public partial class LobbyManager
         catch (Exception e)
         {
             Logger.Error($"Failed to get online friends. {e.Message}");
-            return Task.FromResult(new List<GlobalTypes.PlayerInvite>());
+            return Task.FromResult(new List<PlayerInvite>());
         }
     }
 }
