+++ Games/Game.cs
@@ -1,3 +1,5 @@
+using GodotPeer2PeerSteamCSharp.Types.Games;
+
 namespace GodotPeer2PeerSteamCSharp.Games;
 
 public interface IGame
@@ -7,7 +9,7 @@ public interface IGame
         get;
     }
 
-    GlobalTypes.GameType GameType
+    GameType GameType
     {
         get;
     }
