+++ Godot-Peer-2-Peer-Steam-CSharp.csproj
@@ -39,8 +39,8 @@
     </ItemGroup>
 
     <ItemGroup>
-        <Folder Include="Core\Lobby\Gates\" />
         <Folder Include="DLL\"/>
+        <Folder Include="Engine\Autoloads\Multiplayer\" />
         <Folder Include="Multiplayer\Services\Steam\"/>
         <Folder Include="Scenes\Transitions\"/>
         <Folder Include="Scripts\UI\MainMenu\"/>
