+++ Types/Games/Information.cs
@@ -1,0 +1,9 @@
+namespace GodotPeer2PeerSteamCSharp.Types.Games;
+
+public enum GameType
+{
+    Duel,
+    TeamDuel,
+    OneVersusMany,
+    FreeForAll
+}
