+++ project.godot
@@ -17,13 +17,11 @@ config/icon="res://icon.svg"
 
 [autoload]
 
-Logger="*res://Core/Logger.cs"
-EventBus="*res://Core/EventBus.cs"
-LobbyManager="*res://Core/Lobby/LobbyManager.cs"
-TransportManager="*res://Core/Transport/TransportManager.cs"
-InputManager="*res://Core/InputManager.cs"
-SceneManager="*res://Core/SceneManager/SceneManager.cs"
-StateManager="*res://Core/StateManager.cs"
+LobbyManager="*res://Engine/Autoloads/Multiplayer/Lobby/LobbyManager.cs"
+TransportManager="*res://Engine/Autoloads/Multiplayer/Transport/TransportManager.cs"
+SceneManager="*res://Engine/SceneManager/SceneManager.cs"
+StateManager="*res://Engine/Autoloads/StateManager.cs"
+InputManager="*res://Engine/Autoloads/InputManager.cs"
 
 [display]
